Cambios en estabilidad de lectura de medidor. Más info de bateria en log

+++ Core/Src/adc_if.c
@@ -182,8 +182,9 @@ uint16_t SYS_GetBatteryLevel(void)
   batteryLevelmV = (uint16_t)((pin_mV * 403U) / 100U);
 
   /* Debug: print ADC internals to help diagnose first-uplink anomalies */
-  APP_LOG(TS_ON, VLEVEL_M, "DEBUG: SYS_GetBatteryLevel: measVref=%lu vdd_mV=%lu raw_pin=%lu pin_mV=%lu batt_mV=%u\r\n",
-    (unsigned long)measVref, (unsigned long)vdd_mV, (unsigned long)raw_pin, (unsigned long)pin_mV, batteryLevelmV);
+  /* Note: use %u and cast to unsigned int to avoid unsupported long specifiers */
+  APP_LOG(TS_ON, VLEVEL_M, "DEBUG: SYS_GetBatteryLevel: measVref=%u vdd_mV=%u raw_pin=%u pin_mV=%u batt_mV=%u\r\n",
+    (unsigned int)measVref, (unsigned int)vdd_mV, (unsigned int)raw_pin, (unsigned int)pin_mV, (unsigned int)batteryLevelmV);
 
   return batteryLevelmV;
 }
